requirements.txt
codetiming==1.4.0
humanfriendly==10.0
matplotlib==3.10.6
numpy==2.3.2
ogb==1.3.6
omegaconf==2.3.0
pandas==2.3.2
pytorch_lightning==2.5.1.post0
PyYAML==6.0.1
PyYAML==6.0.2
rich==14.1.0
rootutils==1.0.7
scikit_learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentence_transformers==2.2.2
torch==2.6.0+cu124
torch_geometric==2.6.1
tqdm==4.66.5
transformers==4.36.2
